+++ lib/truffle/providers/google.rb
@@ -78,7 +78,7 @@ module Truffle
           error_message: error_message
         )
       rescue Providers::Error => e
-        error_response(e.message, model: request_model)
+        error_response(e.message, model: request_model, retry_after_ms: e.retry_after_ms)
       end
 
       # Streaming counterpart to #chat. Opens an SSE request to
@@ -338,7 +338,8 @@ module Truffle
 
         response = http.request(request)
         unless response.is_a?(Net::HTTPSuccess)
-          raise Error, "Google #{response.code}: #{truncate(response.body)}"
+          raise Error.new("Google #{response.code}: #{truncate(response.body)}",
+                          retry_after_ms: retry_after_ms(response))
         end
 
         JSON.parse(response.body)